+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 Below, you can find all releases of this expansion and their changes.
 
+## [1.2.3]
+**This version is ONLY compatibe with PlaceholderAPI onwards!**
+
+- Updated to PlaceholderAPI 2.10.10 to use `getBoolean(String, boolean)`
+
 ## [1.2.2]
 - Math-Expansion now logs the used placeholder and mentions a proper cause whenever it fails to parse a Math-Expression, precision number or placeholder in general.
 
@@ -61,3 +66,4 @@ This version still has `1.0.7` defined as expansion version. Please update to [t
 [1.2.0]: https://api.extendedclip.com/expansions/math/versions/math-120
 [1.2.1]: https://api.extendedclip.com/expansions/math/versions/math-121
 [1.2.2]: https://api.extendedclip.com/expansions/math/versions/math-122
+[1.2.3]: https://api.extendedclip.com/expansions/math/versions/math-123